refactor: centralize pre-signed input data generation for blockchain services
