--- frontend/src/chain/maya/config.ts
@@ -1,3 +1,0 @@
-export const mayaConfig = {
-  fee: BigInt(2000000000),
-};
