+++ frontend/src/chain/tx/fee/utils/getFeeAmount.ts
@@ -1,6 +1,7 @@
 import { matchDiscriminatedUnion } from '../../../../lib/utils/matchDiscriminatedUnion';
+import { Chain } from '../../../../model/chain';
+import { cosmosGasLimitRecord } from '../../../cosmos/cosmosGasLimitRecord';
 import { KeysignChainSpecific } from '../../../keysign/KeysignChainSpecific';
-import { mayaConfig } from '../../../maya/config';
 import { polkadotConfig } from '../../../polkadot/config';
 import { rippleConfig } from '../../../ripple/config';
 import { tonConfig } from '../../../ton/config';
@@ -13,7 +14,7 @@ export const getFeeAmount = (chainSpecific: KeysignChainSpecific): bigint =>
     suicheSpecific: ({ referenceGasPrice }) => BigInt(referenceGasPrice),
     solanaSpecific: ({ priorityFee }) => BigInt(priorityFee),
     thorchainSpecific: ({ fee }) => BigInt(fee),
-    mayaSpecific: () => mayaConfig.fee,
+    mayaSpecific: () => BigInt(cosmosGasLimitRecord[Chain.MayaChain]),
     cosmosSpecific: ({ gas }) => BigInt(gas),
     polkadotSpecific: () => polkadotConfig.fee,
     tonSpecific: () => tonConfig.fee,
