+++ frontend/src/services/Blockchain/IBlockchainService.ts
@@ -1,11 +1,8 @@
 import { PublicKey } from '@trustwallet/wallet-core/dist/src/wallet-core';
 
 import { tss } from '../../../wailsjs/go/models';
-import { KeysignPayload } from '../../gen/vultisig/keysign/v1/keysign_message_pb';
 
 export interface IBlockchainService {
-  getPreSignedInputData(keysignPayload: KeysignPayload): Promise<Uint8Array>;
-
   executeTransaction(
     publicKey: PublicKey,
     txInputData: Uint8Array,
